backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
Pillow>=9.5.0
transformers>=4.30.0
torch>=2.7.0
torchvision>=0.22.0
python-dotenv==1.0.0
